freq not working

+++ Sources/DOR.h
@@ -35,7 +35,7 @@ typedef struct ChannelThreadData
   OS_ECB* semaphore;
   uint8_t channelNb;
   float irms;
-  float samples[16];
+  int16_t samples[16];
   int16_t sample;
   bool timerStatus;
   uint32_t currentTimeCount;
